@maumar 2024-11-26
This change doesn't have significant impact on AOT scenario

before the fix
Benchmark 1: .\AdventureworksAotEF9.exe
  Time (mean ± σ):     818.8 ms ±  10.9 ms    [User: 402.0 ms, System: 42.2 ms]
  Range (min … max):   803.9 ms … 864.1 ms    100 runs
  
after the fix  
  Benchmark 1: .\AdventureworksAotEF9.exe
  Time (mean ± σ):     819.3 ms ±  14.0 ms    [User: 385.0 ms, System: 44.6 ms]
  Range (min … max):   806.7 ms … 906.6 ms    100 runs